ports/unix: Fix "‘signal_sco_can_send_now’ defined but not used".
Added a "ports/unix/btstack_config.h" for a workaround:
If "MICROPY_BLUETOOTH_BTSTACK_USB" is set and "ENABLE_SCO_OVER_HCI"
is not defined there is a compiler error:
    ../../lib/btstack/platform/libusb/hci_transport_h2_libusb.c:1354:13:
    error: ‘signal_sco_can_send_now’ defined but not used
Because the problem is in an external code a workaround is done by
defining MICROPY_BLUETOOTH_BTSTACK_USB in the added "btstack_config.h".

The "ports/unix/Makefile" has been changed so that the "btstack_config.h"
will be included.

Signed-off-by: Yasmin Bosch <161633073+yasnim24@users.noreply.github.com>